db/requirements.txt
paho-mqtt
pymongo
python-dotenv
